Added Link to Homework folder (#58)
* Update homework.md

Adding the link to the homework folder for easy accessibility. 
I took some time to search around the repo where to find the home folder. As this is the first mention of the homework folder in the entire Doc. That is the reason it is necessary only at the mentioned line

* Update homework.md

The homework link in the line 23 was linked to my own repo (Mekong-delta). Changed it to the link of the DataTalks course's homework folder .